requirements.txt
amqp==2.5.0
asn1crypto==0.24.0
aws-shell==0.2.1
awscli==1.16.69
billiard==3.6.0.0
boto3==1.9.59
botocore==1.12.59
celery==4.3.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
configobj==5.0.6
cryptography==2.3.1
Django==1.11.8
django-cors-headers==2.5.0
django-extensions==2.1.3
django-guardian==1.5.1
django-haystack==2.8.1
django-redis==4.7.0
django-smtp-ssl==1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
docutils==0.14
future==0.17.1
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
jmespath==0.9.3
jpush==3.3.3
kombu==4.6.3
Pillow==7.0.0
prompt-toolkit==1.0.15
pyasn1==0.4.4
pycparser==2.19
Pygments==2.3.0
PyJWT==1.6.4
PyMySQL==0.8.0
pyOpenSSL==18.0.0
python-dateutil==2.7.5
pytz==2018.5
PyYAML==3.13
redis==3.2.0
requests==2.20.0
requests-toolbelt==0.9.1
rsa==3.4.2
s3transfer==0.1.13
serializers==0.2.4
simplejson==3.16.0
six==1.11.0
tqdm==4.45.0
urllib3==1.24
vine==1.3.0
wcwidth==0.1.7
